requirements.txt
flask>=2.3.2
mysql-connector-python>=8.0.33
opencv-python>=4.7.0.72
numpy>=1.24.3
mediapipe>=0.10.0
scipy>=1.10.1
flask-cors>=3.0.10
keras-facenet>=0.3.2
tensorflow>=2.12.0
requests>=2.31.0
mysqlclient>=2.2.0